Add in a new page where an admin can go and edit existing jobs,
The page is accessible through the edit button on each job tile. The edit page will open with text input and text areas filled in with the selected job's details. Exit and save buttons exist to redirect to the dashboard and other pages. Text blocks can be edited, but no changes will be saved to the sample array.